Fixed docopt_c to work in python 3.
There was only one problem, really. Python 3 doesn't understand the long
keyword, since in python 3, longs are the same as ints. I added an if
statement to detect which version of python is being used. It then
either expects long or not as appropriate.

I tested this change by following the example, and everything seems to
work.