Add support for variable prefix
It is currently not possible to use an option or command name that happesn to be
a C keyword. So for example trying to have a command named "register" will fail
with a compilation error.

This introduces support for setting a `prefix` for the variable names defined
in the struct. So for example using prefix `_` would mean you access
the command via `args._register` instead.